refactor:Adjust xds rule build
